+++ src/js/service/YoutubeSearchAPI.js
@@ -22,7 +22,7 @@ export default class YoutubeSearchAPI {
   }
 
   getURL(searchKeyword, nextPageToken) {
-    const url = new URL('dummy/youtube/v3/search', 'https://onstar.netlify.app');
+    const url = new URL('youtube/v3/search', 'https://onstar.netlify.app');
     const parameter = new URLSearchParams({
       part: 'snippet',
       maxResults: 10,
