+++ src/js/validate.js
@@ -1,7 +1,16 @@
-import { ERROR } from './constants/constants.js';
+import { ERROR, MAX_SAVED_VIDEOS_NUMBER } from './constants/constants';
 
 export const validateInput = input => {
   if (!input) {
     throw new Error(ERROR.MESSAGE.EMPTY_INPUT);
+  }
+};
+
+export const validateAbleToSaveVideo = (savedVideos, saveTargetVideo) => {
+  if (savedVideos.length > MAX_SAVED_VIDEOS_NUMBER) {
+    throw new Error(ERROR.MESSAGE.OVER_MAX_SAVED_VIDEOS_NUMBER);
+  }
+  if (savedVideos.some(savedVideo => savedVideo.id === saveTargetVideo.id)) {
+    throw new Error(ERROR.MESSAGE.ALREADY_SAVED_VIDEOS);
   }
 };
